Refactor metadata schema helper (#13)
